fixed rest endpoints for game fixed variable types added name to texasuser

+++ src/main/java/com/aquino/TexasWebService/texas/TexasGame.java
@@ -65,7 +65,7 @@ public class TexasGame implements CardGame {
     }
     
     @Override
-    public User removeUser(int userId) {
+    public User removeUser(long userId) {
         foldUser(userList.get(userId));
         return userList.remove(userId);
     }
